ai_server/requirements.txt
langchain-core
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
langchain
huggingface-hub
python-dotenv
numpy
scikit-learn
sentencepiece
pypdf
chromadb
langchain_community
fastapi
uvicorn